统一文件下载请求方式
